+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 
 - Apply heuristic to suggest using JSX fragments where we guess that might be what the user wanted. https://github.com/rescript-lang/rescript/pull/7714
 - Show deprecation warnings for `bs-dependencies` etc. for local dependencies only. https://github.com/rescript-lang/rescript/pull/7724
+- Add check for minimum required node version. https://github.com/rescript-lang/rescript/pull/7723
 
 #### :bug: Bug fix